services/billings/requirements.txt
Flask==3.1.1
Flask-Cors==6.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.4.0
SQLAlchemy==2.0.42
stripe==13.0.1
requests==2.32.3
python-dotenv==1.0.1